refactor(views): extract remaining computed view properties into View structs
